リファクタリング
